requirements.txt
Cython==3.0.12
Django==5.1a1
django_select2==8.4.0
face_recognition==1.3.0
facenet_pytorch==2.6.0
insightface==0.7.3
keras_facenet==0.3.2
mediapipe==0.10.21
numpy==2.2.5
opencv_contrib_python==4.9.0.80
opencv_python==4.11.0.86
pandas==2.2.3
Pillow==11.2.1
scipy==1.15.2
tensorflow==2.19.0
torch==2.7.0+cu128
torchvision==0.22.0+cu128
ultralytics==8.3.127
